cleaned up file structure, add FindGamePage component skeleton
