+++ client/src/components/ChaseMode.tsx
