+++ client/src/components/KillMode.tsx
