+++ client/src/components/pages/GamePage.tsx
@@ -1,8 +1,8 @@
 import React, {useState} from 'react';
-import KillMode from '../components/KillMode'
+import KillMode from '../KillMode'
 // import ChaseMode from '../components/ChaseMode'
 
-import ChaseCam from '../components/ChaseCam';
+import ChaseCam from '../ChaseCam';
 
 function GamePage() {
   const [gameMode, setGameMode] = useState('Chase')
